<feat>[#283-WHS] 1001

See merge request s09-fintech-finance-sub2/S09P22C109!116

+++ back-end-app/src/main/java/project/app/c109/backendapp/product/domain/entity/Product.java
@@ -35,7 +35,7 @@ public class Product {
 
     // 할인율
     @Column(name = "product_dcrate")
-    private Integer productDcrate;
+    private Double productDcrate;
 
     // 상품설명, null을 허용하지 않고 길이는 최대 2000
     @Column(name = "product_info", nullable = false, length = 2000)
@@ -91,11 +91,11 @@ public class Product {
         this.productPrice = productPrice;
     }
 
-    public Integer getProductDcrate() {
+    public Double getProductDcrate() {
         return productDcrate;
     }
 
-    public void setProductDcrate(Integer productDcrate) {
+    public void setProductDcrate(Double productDcrate) {
         this.productDcrate = productDcrate;
     }
